# Conflicts:
#   backend/src/main/java/com/yourseason/backend/member/consultant/service/ConsultantService.java

+++ backend/src/main/java/com/yourseason/backend/member/customer/controller/dto/ReviewListResponse.java
@@ -13,7 +13,7 @@ public class ReviewListResponse {
     private Long reviewId;
     private String consultantNickname;
     private String consultantImageUrl;
-    private double star;
+    private int star;
     private String comment;
     private LocalDate reviewDate;
 }
